refactor: Split mintThenList script into mint and list scripts.
